load gt masks as int16

+++ viscy/light/data.py
@@ -255,7 +255,7 @@ class MaskTestDataset(SlidingWindowDataset):
         if img_path := self.masks.get(
             (position_name, t_idx, z_idx + self.z_window_size // 2)
         ):
-            sample["mask"] = torch.from_numpy(imread(img_path))
+            sample["mask"] = torch.from_numpy(imread(img_path).astype(np.int16))
         return sample
